+++ internal/lsp/testutil.go
@@ -48,8 +48,15 @@ type lspTester struct {
 func (lt *lspTester) initialize() {
 	resp, serverState := sendAndReceive[lsp.InitializeResult](lt, "initialize", InitializeParams{
 		Capabilities: ClientCapabilities{
-			Diagnostics: DiagnosticWorkspaceClientCapabilities{
-				RefreshSupport: true,
+			TextDocument: &TextDocumentClientCapabilities{
+				Diagnostic: &DiagnosticClientCapabilities{
+					DynamicRegistration: true,
+				},
+			},
+			Workspace: &WorkspaceClientCapabilities{
+				Diagnostics: &DiagnosticWorkspaceClientCapabilities{
+					RefreshSupport: true,
+				},
 			},
 		},
 	})